separate models into specific modules
