+++ qiskit/qobj/__init__.py
@@ -7,14 +7,19 @@
 
 """Module for the Qobj structure."""
 
-from .qobj import Qobj, QasmQobj, PulseQobj
-from .models import (QobjConfig, QobjExperiment, QobjInstruction, QobjHeader,
-                     QobjExperimentConfig, QobjExperimentHeader,
-                     QobjConditional, QobjPulseLibrary, QobjMeasurementOption,
-                     QasmQobjConfig, QasmQobjExperiment, QasmQobjInstruction,
-                     QasmQobjExperimentConfig,
-                     PulseQobjConfig, PulseQobjExperiment, PulseQobjInstruction,
-                     PulseQobjExperimentConfig)
-from .exceptions import QobjValidationError
+from qiskit.qobj.models.base import (QobjInstruction, QobjExperimentHeader, QobjExperimentConfig,
+                                     QobjExperiment, QobjConfig, QobjHeader)
+
+from qiskit.qobj.models.pulse import (PulseQobjInstruction, PulseQobjExperimentConfig,
+                                      PulseQobjExperiment, PulseQobjConfig,
+                                      QobjMeasurementOption, QobjPulseLibrary)
+
+from qiskit.qobj.models.qasm import (QasmQobjInstruction, QasmQobjExperimentConfig,
+                                     QasmQobjExperiment, QasmQobjConfig,
+                                     QobjConditional)
 
 from ._validation import validate_qobj_against_schema
+
+from .exceptions import QobjValidationError
+
+from .qobj import Qobj, QasmQobj, PulseQobj
